add an Altivec version of adler32
Coded in intrinsic as found in the Altivec PIM, this code
should work with any PPC compiler which knows Altivec.
Only due to lack of testing with other compiler this code
is restricted to GCC ATM.

The code is automatically choosen when the compiler has the
right predefines.

This is not only interesting for PowerPC G4 & G5, but also
the IBM POWER6 and POWER7 which again feature an Altivec
unit, only now it's called VSX.

Here some numbers from an PPC G4 Powerbook:
-------- orig ------
    a: 0x0CB4B676, 10000 * 160000 bytes   t: 22200 ms
    a: 0x25BEB273, 10000 * 159999 bytes   t: 20000 ms
    a: 0x733CB174, 10000 * 159998 bytes   t: 20800 ms
    a: 0x1144AF76, 10000 * 159996 bytes   t: 21200 ms
    a: 0x3F4ECB8A, 10000 * 159992 bytes  t: 21200 ms
    a: 0x1902A382, 10000 * 159984 bytes  t: 21100 ms
-------- altivec ------
    a: 0x0CB4B676, 10000 * 160000 bytes  t: 3400 ms
    a: 0x25BEB273, 10000 * 159999 bytes  t: 3400 ms
    a: 0x733CB174, 10000 * 159998 bytes  t: 3300 ms
    a: 0x1144AF76, 10000 * 159996 bytes  t: 3400 ms
    a: 0x3F4ECB8A, 10000 * 159992 bytes  t: 3400 ms
    a: 0x1902A382, 10000 * 159984 bytes  t: 3300 ms
speedup: 6.52941

The raw engine speed is insane, it is the bigest speedup of all
SIMD versions of adler32 i have written.

Unfortunately the FSB/Memory can't keep up:
-------- orig ------
    a: 0x01A71FA6, 100 * 16000000 bytes  t: 47600 ms
    a: 0x2DEB1BA3, 100 * 15999999 bytes  t: 46100 ms
    a: 0x12481AA4, 100 * 15999998 bytes  t: 47500 ms
    a: 0xDDF018A6, 100 * 15999996 bytes  t: 47600 ms
    a: 0xC43634BA, 100 * 15999992 bytes  t: 47500 ms
    a: 0xF7D70CB2, 100 * 15999984 bytes  t: 47500 ms
-------- altivec ------
    a: 0x01A71FA6, 100 * 16000000 bytes  t: 36000 ms
    a: 0x2DEB1BA3, 100 * 15999999 bytes  t: 36000 ms
    a: 0x12481AA4, 100 * 15999998 bytes  t: 36000 ms
    a: 0xDDF018A6, 100 * 15999996 bytes  t: 36000 ms
    a: 0xC43634BA, 100 * 15999992 bytes  t: 35900 ms
    a: 0xF7D70CB2, 100 * 15999984 bytes  t: 36000 ms
speedup: 1.3222

Still we can squeeze some cycles.