add special adler32 versions for ARM
This adds an NEON version, a iWMMXt version for Intel (now
Marvel) XScale and a version for ARMv6 DSP instructions of
adler32.

The NEON Version is coded in intrinsic. If i understand
things right it should work with any ARM compiler which
understands NEON.

The iWMMXt version tries to use intrinsic, but that is a
little complicated. This way there may be a chance it may
work with the MS Visual Studio for ARM (WinCE), but accord.
to MSDN there seem to be some little discrepancies. To make
matters worse GCC does not like it's own mmintrin.h.
Since i have no XScale avail. the iWMMXt version is untested,
but it is very similar to the x86 MMX version, it should
work(TM) (famous last words...).

The ARMv6 DSP version uses inline ASM, so i restricted it to
GCC (or compatible compiler, like clang, can also be used.
Clang should find the way in by itself, it disguises as
GCC 4.2).

The code is automatically choosen when the right compiler
predefines are present.

Some results:
an i.MX515@800MHz  (arm7l) with NEON
    -------- orig ------
           a: 0x0CB4B676, 10000 * 160000 bytes     t: 4010 ms
           a: 0x25BEB273, 10000 * 159999 bytes     t: 2990 ms
           a: 0x733CB174, 10000 * 159998 bytes     t: 4060 ms
           a: 0x1144AF76, 10000 * 159996 bytes     t: 4050 ms
           a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 4060 ms
           a: 0x1902A382, 10000 * 159984 bytes     t: 4060 ms
    -------- vec ------
           a: 0x0CB4B676, 10000 * 160000 bytes     t: 1450 ms
           a: 0x25BEB273, 10000 * 159999 bytes     t: 1450 ms
           a: 0x733CB174, 10000 * 159998 bytes     t: 1460 ms
           a: 0x1144AF76, 10000 * 159996 bytes     t: 1450 ms
           a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 1460 ms
           a: 0x1902A382, 10000 * 159984 bytes     t: 1450 ms
    speedup: 2.765517
Independent mesurements of some members of the linaro Project
confirm this on another Cortex-A9 & Cortex-A8, with some
variance depending on L2 cache structure.

The bad news is ARMv6 DSP:
an i.MX515@800MHz  (arm7l) with ARMv6 DSP
    -------- orig ------
           a: 0x0CB4B676, 10000 * 160000 bytes     t: 4040 ms
           a: 0x25BEB273, 10000 * 159999 bytes     t: 3880 ms
           a: 0x733CB174, 10000 * 159998 bytes     t: 4070 ms
           a: 0x1144AF76, 10000 * 159996 bytes     t: 4060 ms
           a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 4050 ms
           a: 0x1902A382, 10000 * 159984 bytes     t: 4060 ms
    -------- vec2 ------
           a: 0x0CB4B676, 10000 * 160000 bytes     t: 4240 ms
           a: 0x25BEB273, 10000 * 159999 bytes     t: 4250 ms
           a: 0x733CB174, 10000 * 159998 bytes     t: 3300 ms
           a: 0x1144AF76, 10000 * 159996 bytes     t: 4240 ms
           a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 4240 ms
           a: 0x1902A382, 10000 * 159984 bytes     t: 4250 ms
    speedup: 0.952830
Unfortunatly it is slower then the generic code, so it disabled
and only kept for reference (maybe someone else finds the right
knob to speed it up).