add an Alpha MAX version of adler32
Alpha also had a little SIMD extention called MAX avail.
since the 21264 aka. ev6 (and later in the 21164pc aka
pca56).
A nice packed mul is missing, but an pixel error instruction
(sum of absolute difference) is avail., and byte-halfword
unpack instructions so it is possible to build something.

Thanks to Mike Frysinger, i could test the code on an Alpha.
And here the numbers:
an Alpha EV68 with MAX used (which means BWX is used for the
original code)
  -------- orig ------
         a: 0x0CB4B676, 10000 * 160000 bytes     t: 3646 ms
         a: 0x25BEB273, 10000 * 159999 bytes     t: 3294 ms
         a: 0x733CB174, 10000 * 159998 bytes     t: 3335 ms
         a: 0x1144AF76, 10000 * 159996 bytes     t: 3400 ms
         a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 3351 ms
         a: 0x1902A382, 10000 * 159984 bytes     t: 3369 ms
  -------- MAX ------
         a: 0x0CB4B676, 10000 * 160000 bytes     t: 1953 ms
         a: 0x25BEB273, 10000 * 159999 bytes     t: 1948 ms
         a: 0x733CB174, 10000 * 159998 bytes     t: 1963 ms
         a: 0x1144AF76, 10000 * 159996 bytes     t: 1962 ms
         a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 1958 ms
         a: 0x1902A382, 10000 * 159984 bytes     t: 1964 ms
  speedup: 1.866871