add mips version of adler32
This contains an adler32 version for >=loongson2f and
the MIPS DSP-ASE.

To use these ISA-Extentions inline ASM or compiler
builtins are used, so this code is for GCC only.
The code is automatically choosen by the right
preprocessor defines.

The code for the DSP-ASE is unfort. untested, but
the loongson code looks good:
on an loongson-2f@800MHz (mips)
  -------- orig ------
         a: 0x0CB4B676, 10000 * 160000 bytes     t: 5180 ms
         a: 0x25BEB273, 10000 * 159999 bytes     t: 5240 ms
         a: 0x733CB174, 10000 * 159998 bytes     t: 5310 ms
         a: 0x1144AF76, 10000 * 159996 bytes     t: 5300 ms
         a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 5640 ms
         a: 0x1902A382, 10000 * 159984 bytes     t: 5590 ms
  -------- vec ------
         a: 0x0CB4B676, 10000 * 160000 bytes     t: 2940 ms
         a: 0x25BEB273, 10000 * 159999 bytes     t: 2940 ms
         a: 0x733CB174, 10000 * 159998 bytes     t: 2940 ms
         a: 0x1144AF76, 10000 * 159996 bytes     t: 2940 ms
         a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 2940 ms
         a: 0x1902A382, 10000 * 159984 bytes     t: 2940 ms
  speedup: 1.761905
Big thanks go to Edwin Török for his efforts and time he put
into fiddling with the code, debugging and trying patches.