add special adler32 version for blackfin
This is a special version of adler32 for the Analog Devices
Blackfin CPU, which is a mix between General Purpose CPU
and DSP with a little SIMD.

The code is written as inline ASM, but GCC is the main
(only?) compiler compiler for these CPU.
In return we get debug info and handling of calling
convetions (pic vs. nopic, elf vs bare) for free.
With the right preprocessor defines the code is
automatically choosen.

Again a big thanks to Mike Frysinger for testing grounds.

I'm a little bit unhappy with the performance, but
ENOREGISTER.
The numbers:
An Blackfin BF537@500MHz
  -------- orig ------
         a: 0x0CB4B676, 10000 * 160000 bytes     t: 19200 ms
         a: 0x25BEB273, 10000 * 159999 bytes     t: 19200 ms
         a: 0x733CB174, 10000 * 159998 bytes     t: 19200 ms
         a: 0x1144AF76, 10000 * 159996 bytes     t: 19190 ms
         a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 24150 ms
         a: 0x1902A382, 10000 * 159984 bytes     t: 24160 ms
  -------- vec ------
         a: 0x0CB4B676, 10000 * 160000 bytes     t: 10690 ms
         a: 0x25BEB273, 10000 * 159999 bytes     t: 10690 ms
         a: 0x733CB174, 10000 * 159998 bytes     t: 10700 ms
         a: 0x1144AF76, 10000 * 159996 bytes     t: 10680 ms
         a: 0x3F4ECB8A, 10000 * 159992 bytes     t: 10690 ms
         a: 0x1902A382, 10000 * 159984 bytes     t: 10670 ms
  speedup: 1.796071

Putting vorder_o/e into different L1 cache banks brings that
to 1.84, but since that eats a precious resource for little
gain, and the compiler can react a little itchy when told to
so in certain configs, i left it out.

And even if i'm unhappy with the speedup, these values seem
to be impressive, since the test works with chunks larger then
the cache, so it totally memory bound.
Sticking to blocks within the cachesize like for example 16k
gives a speedup of 2.25.