change x86 SSSE3 version to use palignr
SSSE 3 did not only add pmaddubsw, an instruction interesting to our
problem, but also palignr.
palignr allows to extract 128 bit byte aligned out of a field of 256
bit. This can be uses to align unaligned data. Hence the name.

Only problem is: The extraction offset is a constant compiled into
the instruction => *facepalm*. As if you would knew the misalignment
beforehand most of the time. Self modifing Code? If you do such stuff,
you have to flush the hell out of all caches and buffers and pipelines,
syncronize all CPUs. Please, it's 2012, memory protection (RO, NX),
multi threading and stuff, really? *sigh*.

So the use of palignr for the task the name suggests is a little
complicated. One has to either create an code path for every
possible missalignment or a helper function which branches to the
right instruction for a dynamic input value, which isn't very fast.
ARM NEON has the same problem.

But lo and behold, since we only have to use the alignment on start
and trailer, with a little C&P coding, helper can be created and the
problem can be overcome, for a nice win in the unaligned case:

         -------- old SSSE3 ------
                a: 0x47D227BB, 10000000 * 84 bytes      t: 2700 ms
                a: 0x27A52450, 10000000 * 83 bytes      t: 3500 ms
                a: 0x1146240B, 10000000 * 82 bytes      t: 3400 ms
                a: 0xC2D62315, 10000000 * 80 bytes      t: 3400 ms
                a: 0x0B2B418F, 10000000 * 76 bytes      t: 3100 ms
                a: 0x2BF71F2C, 10000000 * 68 bytes      t: 3400 ms
         -------- new SSSE3 ------
                a: 0x47D227BB, 10000000 * 84 bytes      t: 2600 ms
                a: 0x27A52450, 10000000 * 83 bytes      t: 2800 ms
                a: 0x1146240B, 10000000 * 82 bytes      t: 2800 ms
                a: 0xC2D62315, 10000000 * 80 bytes      t: 2800 ms
                a: 0x0B2B418F, 10000000 * 76 bytes      t: 2900 ms
                a: 0x2BF71F2C, 10000000 * 68 bytes      t: 2800 ms

This esp. helps on short length in the unaligned case.