overhaul x85 adler32 inner SSSE3 loop a little
pmaddubsw has one downside: it only widens from bytes to words.
For that reason a tight inner loop has to be run, with max. 8
iterations (and unrolling the loop is not very beneficial...).

But we can optimize the loop setup a little bit, by precalculating
the number of full 8 iteration runs, now that one more ALU register
is used anyway.
This gives a slight benfit of some percent.